requirements.txt
discord.py[voice]
discord.py
PyMySQL
requests
ffmpeg
discord-timers
numpy
wikipedia-api
